requirements.txt
opencv-python
Pillow
numpy
flask
tensorflow
base64